requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
chromadb
transformers[torch]
torch
kafka-python
pydantic
pytest
